+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Future
 
 ---
 
-* STAR-624: Updated SensCalc mapping for LOW ( Zoom & Continuum )
+* STAR-624: Updated SensCalc mapping & display ( Zoom & Continuum )
 * STAR-536: Ability to Edit an Observation
 * STAR-609: Observation updates in regards to MID Telescope
 * STAR-615: Proposal mapping: